Simplify display rules for show-children icon
We were using `display: inline` assuming that's the way elements will be
displayed by default when shown. However, those elements could be
displayed in a different way (inline-flex, for instance). So we avoid
any possible conflicts by using the `display: none` rule when we want to
hide the elements.

Besides, the code is now symmetrical and IMHO easier to follow.